fix hooks build
