--- packages/cloud-react/lib/hooks.tsx
@@ -1,24 +1,0 @@
-/* @license Copyright 2023 @paritytech/polkadot-cloud authors & contributors
-SPDX-License-Identifier: GPL-3.0-only */
-
-// TODO: this file should be generated on build.
-
-import { ExtensionsContext } from "./connect/ExtensionsProvider";
-import { OverlayContext } from "./overlay/OverlayProvider";
-import { AnyJson, AnyFunction } from "./utils/types";
-import { useContext, useEffect, useRef } from "react";
-
-export const useOverlay = () => useContext(OverlayContext);
-
-export const useExtensions = () => useContext(ExtensionsContext);
-
-export const useEffectIgnoreInitial = (fn: AnyFunction, deps: AnyJson[]) => {
-  const isInitial = useRef<boolean>(true);
-  useEffect(
-    () => {
-      if (!isInitial.current) fn();
-      isInitial.current = false;
-    },
-    deps ? [...deps] : undefined
-  );
-};
