+++ packages/cloud-react/scripts/prebuild.mjs
@@ -9,7 +9,7 @@ const main = async () => {
   generateTypeExports("./lib", "./lib/types/index.ts");
 
   // Generate entry files.
-  await generateExportEntries({ ignore: ["styles", "svg", "utils", "hooks"] });
+  await generateExportEntries({ ignore: ["styles", "svg", "utils"] });
 };
 
 await main();
